server/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
asyncpg
databases
pydantic>=2.4.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
pydantic[email]>=2.4.0
python-multipart>=0.0.5